ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.41.0
pydantic>=2.12.0
python-multipart>=0.0.20
python-dotenv>=1.0.1
httpx>=0.28.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
bs4>=0.0.1
numpy>=1.26.0
scipy>=1.11.0
semgrep>=1.90.0
click>=8.1.8
tomli>=2.0.1
anthropic>=0.25.0
openai>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic-settings>=2.0.0
python-dateutil>=2.8.2
reportlab>=4.0.0
sqlalchemy>=2.0.0
